order/requirements.txt
asgiref==3.4.1
black==21.12b0
click==8.0.3
Django==3.1.3
django-cors-headers==3.5.0
django-mysql==3.9.0
djangorestframework==3.13.1
mypy-extensions==0.4.3
mysqlclient==2.0.1
pathspec==0.9.0
pika==1.1.0
platformdirs==2.4.0
PyJWT==2.3.0
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
tomli==1.2.3
typing_extensions==4.0.1
